+++ Assets/Pixels/Materials/GreenUp.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 376f63b75300a754c8acfe0770f8043b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
