+++ Assets/Resources/Prefabs/Minimap.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6b5b6f12a215fc54da385507b958ca63
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
