+++ Assets/Resources/Prefabs/Minimap/GreenUp.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 508642f6564ab54478075138c9bde6c6
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
