+++ Assets/Resources/Prefabs/Puzzle/SwordTrap.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9f939698a596f3a44bb748466323b116
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
