+++ Assets/Scripts/Obstacles/SpikeBalls/BallBehavior.cs
@@ -15,51 +15,12 @@ public class BallBehavior : MonoBehaviour {
     // Start is called before the first frame update
     void Start()
     {
-        
+        Destroy(gameObject, 15f);
     }
 
     // Update is called once per frame
     void Update()
     {
-        isGrounded = Physics2D.OverlapCircle(transform.position, collisonRadius, whatIsGround);
-        FSM();
     }   
 
-    void FSM() {
-        switch(mState) {
-            case BallBehaviorState.LowerState:
-                LowerState();
-                break;
-            case BallBehaviorState.RiseState:
-                RiseState();
-                break;
-        }
-    }
-
-    void RiseState() {
-        Vector3 curr = transform.position;
-        curr.y += speed * Time.deltaTime;
-        if (isGrounded) {
-            mState = BallBehaviorState.LowerState;
-            curr.y -= 2 * speed * Time.deltaTime;
-            transform.position = curr;
-        } else {
-            transform.position = curr;
-        }
-    }
-
-    void LowerState() {
-        Vector3 curr = transform.position;
-        curr.y -= speed * Time.deltaTime;
-        if (isGrounded) {
-            mState = BallBehaviorState.RiseState;
-            curr.y += 2* speed * Time.deltaTime;
-            transform.position = curr;
-        } else {
-            transform.position = curr;
-        }
-    }
-    private void OnDrawGizmos() {
-        Gizmos.DrawWireSphere(transform.position, collisonRadius);
-    }
 }
